images/safari_vlm_train/requirements.txt
transformers>=4.57.0
peft>=0.13.0
trl>=0.12.0
qwen-vl-utils
datasets
huggingface_hub
accelerate
Pillow
requests
pydantic>=2.0
prefect>=3.0
wandb